chore(typescript): migrated `victory-label` to typescript
